wip: js/app.js cleanup
