+++ elixir/web-ng/assets/js/lib/god_view/rendering_graph_layer_methods.js
@@ -1,209 +1,30 @@
-import {COORDINATE_SYSTEM} from "@deck.gl/core"
-import {ArcLayer, LineLayer, ScatterplotLayer, TextLayer} from "@deck.gl/layers"
+import {godViewRenderingGraphLayerNodeMethods} from "./rendering_graph_layer_node_methods"
+import {godViewRenderingGraphLayerTransportMethods} from "./rendering_graph_layer_transport_methods"
 
-import PacketFlowLayer from "../deckgl/PacketFlowLayer"
-
-export const godViewRenderingGraphLayerMethods = {
+const godViewRenderingGraphLayerCoreMethods = {
   buildGraphLayers(effective, nodeData, edgeData, edgeLabelData) {
-    const pulse = (Math.sin(this.animationPhase * 3.5) + 1) / 2
-    const pulseRadius = 14 + pulse * 20
-    const pulseAlpha = Math.floor(80 + pulse * 130)
-    const rootPulseNodes = nodeData.filter((d) => d.state === 0)
-    const packetFlowData = this.buildPacketFlowInstances(edgeData)
-
-    const mantleLayers = this.layers.mantle
-      ? [
-          new LineLayer({
-            id: "god-view-edges-mantle",
-            data: edgeData,
-            coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-            getSourcePosition: (d) => d.sourcePosition,
-            getTargetPosition: (d) => d.targetPosition,
-            getColor: (d) => this.edgeTelemetryColor(d.flowBps, d.capacityBps, d.flowPps, false),
-            getWidth: (d) => this.edgeWidthPixels(d.capacityBps, d.flowPps, d.flowBps) + (this.edgeIsFocused(d) ? 1.25 : 0),
-            widthUnits: "pixels",
-            widthMinPixels: 1,
-            pickable: true,
-            parameters: {
-              blend: true,
-              blendFunc: [770, 1, 1, 1],
-              depthTest: false,
-            },
-          }),
-        ]
-      : []
-
-    const crustLayers =
-      this.layers.crust
-        ? [
-            new ArcLayer({
-              id: "god-view-edges-crust",
-              data: edgeData,
-              coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-              getSourcePosition: (d) => [d.sourcePosition[0], d.sourcePosition[1], 8],
-              getTargetPosition: (d) => [d.targetPosition[0], d.targetPosition[1], 8],
-              getSourceColor: (d) => this.edgeTelemetryArcColors(d.flowBps, d.capacityBps, d.flowPps).source,
-              getTargetColor: (d) => this.edgeTelemetryArcColors(d.flowBps, d.capacityBps, d.flowPps).target,
-              getWidth: (d) => {
-                const base = Math.max(1.1, Math.min(this.edgeWidthPixels(d.capacityBps, d.flowPps, d.flowBps) * 0.85, 4.8))
-                return this.edgeIsFocused(d) ? Math.min(5.8, base + 0.9) : base
-              },
-              widthUnits: "pixels",
-              greatCircle: false,
-              getTilt: effective.shape === "local" ? 16 : 24,
-              pickable: true,
-              parameters: {
-                blend: true,
-                blendFunc: [770, 1, 1, 1],
-              },
-            }),
-          ]
-        : []
-
-    const atmosphereLayers = this.layers.atmosphere
-      ? [
-          new PacketFlowLayer({
-            id: "god-view-atmosphere-particles",
-            data: packetFlowData,
-            coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-            pickable: false,
-            time: this.animationPhase,
-            parameters: {
-              blend: true,
-              blendFunc: [770, 1, 1, 1],
-              depthTest: false,
-            },
-          }),
-        ]
-      : []
-
-    const securityLayers = this.layers.security
-      ? [
-          new ScatterplotLayer({
-            id: "god-view-security-pulse",
-            data: rootPulseNodes,
-            coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-            getPosition: (d) => d.position,
-            getRadius: pulseRadius,
-            radiusUnits: "pixels",
-            radiusMinPixels: 8,
-            filled: false,
-            stroked: true,
-            lineWidthUnits: "pixels",
-            getLineWidth: 2,
-            getLineColor: [
-              this.visual.pulse[0],
-              this.visual.pulse[1],
-              this.visual.pulse[2],
-              pulseAlpha,
-            ],
-            pickable: false,
-          }),
-        ]
-      : []
-
-    const baseGeoLines = this.geoGridData()
-    const baseLayers = baseGeoLines.length > 0
-      ? [
-          new LineLayer({
-            id: "god-view-geo-grid",
-            data: baseGeoLines,
-            coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-            getSourcePosition: (d) => d.sourcePosition,
-            getTargetPosition: (d) => d.targetPosition,
-            getColor: [32, 62, 88, 65],
-            getWidth: 1,
-            widthUnits: "pixels",
-            pickable: false,
-          }),
-        ]
-      : []
+    const {
+      baseLayers,
+      mantleLayers,
+      crustLayers,
+      atmosphereLayers,
+      securityLayers,
+    } = this.buildTransportAndEffectLayers(effective, nodeData, edgeData)
 
     return [
       ...baseLayers,
       ...mantleLayers,
       ...crustLayers,
-      new ScatterplotLayer({
-        id: "god-view-nodes",
-        data: nodeData,
-        coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-        getPosition: (d) => d.position,
-        getRadius: (d) => Math.min(8 + ((d.clusterCount || 1) - 1) * 0.45, 26),
-        radiusUnits: "pixels",
-        radiusMinPixels: 4,
-        stroked: true,
-        filled: true,
-        lineWidthUnits: "pixels",
-        pickable: true,
-        getLineWidth: (d) => (d.selected ? 3 : 1),
-        getLineColor: [15, 23, 42, 255],
-        getFillColor: (d) => (this.layers.security ? this.nodeColor(d.state) : this.nodeNeutralColor(d.operUp)),
-      }),
+      ...this.buildNodeAndLabelLayers(effective, nodeData, edgeLabelData),
       ...securityLayers,
       ...atmosphereLayers,
-      ...(this.layers.mantle && (effective.shape === "local" || effective.shape === "regional" || effective.shape === "global")
-        ? [
-            new TextLayer({
-              id: "god-view-node-labels",
-              data: nodeData,
-              coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-              getPosition: (d) => d.position,
-              getText: (d) => d.label,
-              getSize: effective.shape === "local" ? 12 : 10,
-              sizeUnits: "pixels",
-              sizeMinPixels: effective.shape === "local" ? 10 : 8,
-              getColor: this.visual.label,
-              getPixelOffset: [0, -16],
-              billboard: true,
-              pickable: false,
-            }),
-            new TextLayer({
-              id: "god-view-node-metrics",
-              data: nodeData,
-              coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-              getPosition: (d) => d.position,
-              getText: (d) => d.metricText,
-              getSize: effective.shape === "local" ? 10 : 9,
-              sizeUnits: "pixels",
-              sizeMinPixels: 8,
-              getColor: [148, 163, 184, 220],
-              getPixelOffset: [0, -3],
-              billboard: true,
-              pickable: false,
-            }),
-            new TextLayer({
-              id: "god-view-node-status-icon",
-              data: nodeData,
-              coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-              getPosition: (d) => d.position,
-              getText: (d) => d.statusIcon,
-              getSize: effective.shape === "local" ? 12 : 11,
-              sizeUnits: "pixels",
-              sizeMinPixels: 9,
-              getColor: (d) => this.nodeStatusColor(d.operUp),
-              getPixelOffset: [-18, -16],
-              billboard: true,
-              pickable: false,
-            }),
-          ]
-        : []),
-      ...(this.layers.mantle && (effective.shape === "local" || effective.shape === "regional")
-        ? [
-            new TextLayer({
-              id: "god-view-edge-labels",
-              data: edgeLabelData,
-              coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
-              getPosition: (d) => d.midpoint,
-              getText: (d) => d.connectionLabel,
-              getSize: 10,
-              sizeUnits: "pixels",
-              sizeMinPixels: 8,
-              getColor: this.visual.edgeLabel,
-              billboard: true,
-              pickable: false,
-            }),
-          ]
-        : []),
     ]
   },
 }
+
+export const godViewRenderingGraphLayerMethods = Object.assign(
+  {},
+  godViewRenderingGraphLayerCoreMethods,
+  godViewRenderingGraphLayerTransportMethods,
+  godViewRenderingGraphLayerNodeMethods,
+)
